test(web): preview の E2E を公開・所有者・履歴の 3 spec へ分け、順序依存を解消する
preview.spec.ts が 576 行で tests の上限 500 行を超えていた。トップレベルの
test.describe 3 つに境界があったので、そのままファイルへ移した。

分割したところ 5 件が落ちた。workers は 1 で並列実行ではなく、原因は実行順の変化。
Playwright はファイル名順に走るため、分割前は「公開プレビュー → 所有者の操作」の
順だったものが「preview-owner → preview」に入れ替わり、リネームと pin 解除が
公開プレビューの期待値（ファイル名・残日数）を先に壊していた。元のファイルは
1 ファイル内の記述順で暗黙にこの依存を満たしていた。

seed.sql が既に「状態を変える試験ごとに行を分ける」方針を宣言している（削除は
E2EDelete001 で分離済み）ので、それに従って破壊的な 2 件へ専用行を足した。
E2ERename001 は E2EReady0001 と、E2EUnpin0001 は E2EPinned001 と同じ相対日数で
入れている（pinnedRemainingDaysAfterUnpin = 27 はこの相対値から出る定数のため）。

screenshotPath は 3 spec すべてが使うので screenshot.ts へ出した。stubHistory /
historyMovie は履歴、boxOf は所有者だけが使うので移動先へ持たせた。

検証: E2E 158 件すべてパス（分割前と同数）、実行順を入れ替えた 2 通りで同じ結果に
なることを確認（preview-owner→preview / preview→preview-owner とも 31 件パス）、
check-architecture の file-too-large 0 件、typecheck 成功

Closes #202
